Update INIT.c
Extension version 0.4 (Phipli branch).

This is the initial commit. Changes are expected soon following code feedback from cheesestraws.

NOTE - Github uses different newline characters to the classic Mac OS, so this code is for viewing online mainly. You'll be best to take the binary version for use on a vintage computer, or be prepared to convert the newlines (for example using BBEdit).